requirements.txt
selenium
pandas
requests
bs4
click
numpy
mysql-connector-python
html5lib
logging
datetime